requirements.txt
absl-py==2.1.0
altair==5.5.0
annotated-types==0.7.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cycler==0.12.1
filelock==3.17.0
flatbuffers==25.1.24
fonttools==4.55.8
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-genai==0.3.0
gspread==6.1.4
httplib2==0.22.0
idna==3.10
jax==0.5.0
jaxlib==0.5.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.20
ml_dtypes==0.5.1
mpmath==1.3.0
narwhals==1.24.0
networkx==3.4.2
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==4.25.6
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.0a2
pydantic_core==2.29.0
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scipy==1.15.1
seaborn==0.13.2
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
sounddevice==0.5.1
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.69
ultralytics-thop==2.0.14
urllib3==2.3.0
watchdog==6.0.0
websockets==14.2
